# apps/api/openapi.yaml
openapi: 3.0.3
info:
  title: FinVault API
  description: Banking simulator API for test automation practice
  version: 1.0.0
servers:
  - url: http://localhost:4000/api
paths:
...
  /auth/logout:
    post:
      tags: [Auth]
      security: [{ bearerAuth: [] }]
  /auth/refresh:
    post:
      tags: [Auth]
      security: [{ bearerAuth: [] }]
...
  /auth/me:
    get:
      tags: [Auth]
      security: [{ bearerAuth: [] }]
    patch:
      tags: [Auth]
      security: [{ bearerAuth: [] }]
  /accounts:
    get:
      tags: [Accounts]
      security: [{ bearerAuth: [] }]
    post:
      tags: [Accounts]
      security: [{ bearerAuth: [] }]
  /accounts/{id}:
    get:
      tags: [Accounts]
      security: [{ bearerAuth: [] }]
    patch:
      tags: [Accounts]
      security: [{ bearerAuth: [] }]
    delete:
      tags: [Accounts]
      security: [{ bearerAuth: [] }]
  /accounts/{id}/transactions:
    get:
      tags: [Accounts]
      security: [{ bearerAuth: [] }]
  /accounts/{id}/balance:
    get:
      tags: [Accounts]
      security: [{ bearerAuth: [] }]
  /accounts/{id}/deposit:
    post:
      tags: [Accounts]
      security: [{ bearerAuth: [] }]
  /accounts/{id}/withdraw:
    post:
      tags: [Accounts]
      security: [{ bearerAuth: [] }]
  /accounts/{id}/close:
    post:
      tags: [Accounts]
      security: [{ bearerAuth: [] }]
...
  /cards:
    get:
      tags: [Cards]
      security: [{ bearerAuth: [] }]
    post:
      tags: [Cards]
      security: [{ bearerAuth: [] }]
  /cards/{id}/freeze:
    post:
      tags: [Cards]
      security: [{ bearerAuth: [] }]
  /cards/{id}/unfreeze:
    post:
      tags: [Cards]
      security: [{ bearerAuth: [] }]
  /cards/{id}/transactions:
    get:
      tags: [Cards]
      security: [{ bearerAuth: [] }]
  /cards/{id}/limits:
    post:
      tags: [Cards]
      security: [{ bearerAuth: [] }]
  /budgets:
    get:
      tags: [Budgets]
      security: [{ bearerAuth: [] }]
    post:
      tags: [Budgets]
      security: [{ bearerAuth: [] }]
  /budgets/{id}/spending:
    get:
      tags: [Budgets]
      security: [{ bearerAuth: [] }]
  /budgets/{id}/alerts:
    post:
      tags: [Budgets]
      security: [{ bearerAuth: [] }]
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT